Removed WpfExtension
